fix: fix refactor form
