+++ packages/client/src/clientPages/questionCreate/ui/ClientQuestionCreatePage.tsx
@@ -1,10 +1,13 @@
 'use client'
 
-import { useCreateQuestionForm } from '@entities/question'
+import {
+   useQuestionForm,
+   QUSETION_FORM_DEFAULT_VALUES,
+} from '@entities/question'
 import { QuestionCreateInputs } from '@widgets/QuestionCreateInputs'
 
 export function ClientQuestionCreatePage() {
-   const form = useCreateQuestionForm()
+   const form = useQuestionForm(QUSETION_FORM_DEFAULT_VALUES)
    // api 연동시 onSubmit 함수 구현
 
    return (
