+++ packages/client/src/entities/question/model/questionFormDefaultValue.ts
@@ -1,0 +1,8 @@
+import { QuestionFormValues } from '../lib/hook/useQuestionCreateForm'
+
+export const QUSETION_FORM_DEFAULT_VALUES: QuestionFormValues = {
+   title: '',
+   content: '',
+   jobCategory: null,
+   reward: 1000,
+}
